Issue #115 Attempt to convert balance to 32-bit value -e Original/build commit ID: 0e73d2c56d5b25c20539992c252612f45e2c6158

+++ endgame.c
@@ -16,7 +16,7 @@
 
 void endgame()
 {
-    UINT16 total_balance = 0;
+    UINT32 total_balance = 0;
     BOOLEAN negative_total = 0;
     UBYTE tile_data;
     UINT8 current_x;
